chore: modernize codebase and tweak generated files
BREAKING CHANGE: This release includes a total overhaul of the shape of the generated code. The shape of the generated code is designed to be more amenable to tree-shaking by putting everything at the top level. There are no exported `Types` and `Codecs` namespaces anymore.

Types and validation functions are exported directly as top-level types and functions.